Fixed reentrant code for setup_sensors() to correctly return without doing anything if called repeatedly
Changed the Wifi connect loop in start_sta_mode()
to wait only 60 seconds per attempt and to
track the number of attempts made. If the
attempts exceeds 5, it reboots the device.

The main loop in STA mode then checks for wifi
connected state and calls start_sta_mode()

This gives us a better approach to the STA
mode in that it will repeatedly try to connect
and reboot every 5 failures.

If at any time during normal operation, the loop
detects a WiFI outage, it will trigger a reconnect
sequence again and we'll reboot if that fails
5 times.

WiFI on the ESP8266 is not the strongest and many
have reported it as somewhat finnicky with losing
connections or failing to connect to some routers
sometimes. So its better to have a continual
reconnect model and a reboot here and there will
also flush out any unworkable states