Added basic threading around the zeroconf discovery to allow it run continuously for 60 seconds, stop and go again. This parallelises the act of populating the URL dictionary of discovered devices.
The main loop then manages the sunset calculation
and polling of known devices every 60 seconds by calculating
intervals since the last device poll.

Its cheap and cheerful but should drastically improve the
behaviour. Prior to this change, the list of devices was
reset prior to each zeroconf discovery. This ended up causing
gaps in the device polling where one or both devices were
not discovered in some of the 60 second zeroconf detections.

Now the background zeroconf stage is adding to that dict as
new devices are discovered. If devices vanish temporarily
or permanently, the 2 second HTTP timeout on the poll will
skip until it returns if at all.