Changed sample time for sensors to use epoch of most recent minute int(epoch) / 60 * 60
That ensures that from an analytics perspective, the
data will show up for each actual minute rather than
be slave to the exact time we log data which is the way
it was working.

That old approach caused some minutes to never appear
because we were sampling around :58/59 and just straddled
into the next minute for the last device causing it to
appear to have no sample for the previous minute where we
had recorded samples for others.

So the timing is looking to sample at gaps of >= 60 seconds
since the last sample and then record based on the epoch
value of the owning minute prior to the sample pulls