Added config reset mode and wifi LED changes
Changed Wifi status LED to be turned off after
we connect to Wifi.

Added a config reset option to the JSON web handler
that will reset config if it receives a "reset" in the
POST or GET args. All this does is commit config to
EEPROM with marker of 0 guaranteeing a reset on next
load