Added baseline for the S20 smart plug device Very similar to the Sonoff basic except the switch does not include a LED i(GPIO13) as it ships with a blue LED tied to the relay (GPIO12).. comes on/off in sync with the relay.
So the green LED (GPIO13) was presented as a separate switch
assigned only with a LED, no relay. Allows for
separate integration to use that LED for some purpose

The Wifi LED however still uses GPIO13 so the initial flashing
indicators for mode etc use this LED in the same fashion

Serial logging is disabled here as we want to use additional
devices but need to repurpose Tx/Rx GPIO pins and as such
have to nobble serial. I may look to smarten this up a bit
and have something that analyses the switches/sensors pin usage
and if it finds serial Tx/Rx pin usage, then nobbles Serial at
boot. More to follow on this