rework of the caching of device URLs and states
Reworked the discovery stage to collect detected
device URLs in a set.

Then the probe stage takes a list copy of this set
and probes each URL, storing its status in a
dictionary keyed on url. The webpage is generated
on this status dictionary data.

We also store a dict of timestamps keyed on url
and this is updated each time we have a successful
probe of a device.

A housekeeping loop then detects stale timestamps
and removes the cachd status and URL from the
related dictionaries.