Rework of device status dict & webpage layout
Reworked the status dict to store the parsed JSON
data instead. Also moved the JSON parsing into the
fetch_url function. The end result is that we fetch a
URL and get served up a JSON dict as the return result.

Redesigned the population of the webpage to focus
on grouping switches and sensors into zones. Makes
for a more logical grouping as original intended
with the zone field.

Updated automated device register to reflect new
naming after I put a bunch of test devices all
under a single Attic zone