Fixed wifi status check watchdog risk
Fixed what I think is a root cause for some devices
that experience watchdog resets. Previously this was
"fixed" by adding watchdog calls but one Sonoff basic
device would get into a loop of resets occasionally.
The issue would go away and then reoccur days later.

I did notice on scrutiny that the device web server
seemed to lag when this was about to occur. However
it seemed related to WiFI as even ping requests would i
fail.

I tried a hunch that the main loop continual check
on wifi status was too agressive and put a simple track
of the millis() function to detect >= 5s intervals
and then do the wifi check. This immediately
sorted the one device experiencing resets. That same
interval check uses unsigned long arithmetic and will
nicely handle the wrap around of the value.

Also changed the post-wifi connect quick flash burst
from 20 to 50 repetitions to make it a little longer
on the eye.