Reworked internal register to store on device_name key
Changed the internal status dict to store on device name for key, not
URL. Then added a device_name to URL dict.

So the discovery stage will populate the set of URLs. The probe stage
will then iterate that set and populate the dict of status values
keyed on device name, the device_url dict, keyed on device name
and the updated timestamp dict, again keyed on device_name.

Then the action URLs passed to the web client pass back
the device name (used to be URL) and the control and desired
state.

The GET action will then resolve name to URL and build
the URL to invoke the desired action.

Stale activity around status will purge from the status dict,
the URL and TS dicts to clean up.