Added callback support for device-driven update
The JSON status GET can now include args update_ip
and update_port to specify a HTTP IP:port for
a POST event when the device wishes to manually
push for a status update.

The POST carries a single arg.. update=name where name
is the device's own name. The receiving server will then
pull a status and update its records.

When manual switches are pressed, the firmware will
now send off this POST to reaquest its status be
updated. This will not happen if the push IP has not
been set