Reworked AP mode and profile selection
Reworked the AP setup to first prompt for
profile and then render the profile config
form.

You can also select a reset combo and set to yes
and then press apply to return to the same profile
selection screen.

This makes the initial config state easier to manage
and more intuitive.