Changed font of timestamp field, sorted zones
Changed the font and colour details on the dashboard time
stamp.

Modified the zone presentation to sort alphabetically.
Also coded up a sort based on number of controls +
sensors to allow for a possible sort based on a best fit
responsive arrangement. Needs more work and thought
on how a batter self-arrange can be accomplished