reworked dashboard layout into N vertical columns
Got the switch action GETs and refresh GET to pass the window width.
This is then divided by the widget width + a grace of 10 pixels.

This gives us a crude number of columns to use. Then when building
the web page content for the dashboard, we are iterating the
population of N entries in a list, each being a logical column
of the overall single row table.