Changed column distribution of widgets to be based on size
Removed the modulo distribution of widgets based on
count and replaced with selecting a column accoridng to
total size of existing columns.

We already had a zone size dictionary based on a sum of controls
and sensors per zone.

As zones are distributed to a column, we track that additional
size on a dictionary keyed on column index.

Then on the next zone distribution, we select the column
with the smallest accumulated size.

The end result is a fairly balanced distribution of zones
across columns