tweaks to table formatting for dashboard
Fixed some lingering column spacing that could take
place if the table tried to occupy to much of the page
width.

So I removed ended up changing the table to assert a 50%
width and removed the asserted column width in each cell.

That sorted the layout to be more of a tight grid now.