Added instant reload of console pages
Added checks to the zone and device views to detect the
width arg as non-present. Then a simple javascript template
is sent down that will instantly refresh the page but now
pass the page width.

This means the first draw will know the correct width and
have a suitable column count.

Tweaked the icons for the JSON and reboot links