Code indet tidy up, reboot all support and Master widget in device console
Tidied up code and indents as best with Python and its awkward
approach to indentation.

Added a reboot all action to the mix that scans through the known
list of devices and then issus reboot directives to each

Added a master dashbox widget in the device console with
the number of devices and an action to reboot all devices.

Finally came to rest on the per-device JSON handling. It
opens a new window on a targetted frame name. I tried a few
variants from grabbing content and updating a textarea to a popup
alert with the data. In the end, the most practical approach is to
just open a window with the raw URL. It lets a debug mode operate in the
most flexible manner with manual refreshes on that URL as required.