Added support for H801 Wifi LED Controller You can address the LED device by control name and set its state similar to a switch.
However the state is a 4-octet int carrying
a Hue, Red, Green and Blue value.

The hue is optional and when set acts as a scaling
factor for the RGB values when applied to the 8266
PWM range of 1023. It's a crude brightness control
that should help give the same colour shade at
different brightness levels

The URL args can also be sent in for LED controls
using hex or decimal format:

Set colour to 5
http://a.b.c.d/json/control=ABC&state=5

Set Hue to 0x34
Set Red to 0x40
Set Green to 0x50
Set Blue to 0x60
http://a.b.c.d/json/control=ABC&state=0x34405060

Each RGB device has a pin for Red, Green & Blue
including an optional manual pin which generates
random colours for the selected device.

To use a single colour LED device, just set the
blue PIN in the register. As such values for red and
green will be ignored.

Alternatively, a single LED device could have three separate
LEDs or strips set to each of the RGB pins, giving a means of
setting a single colour value that is essentially three
settings in one.

Fading of colour from the current to new value is also possible
when using the delay arg..

http://a.b.c.d/json/control=ABC&state=0x34405060&delay=300

The mechanism uses the main loop to drive the gradual fading of
the RGB values from their current to esired setting. The delay
setting is a modulo mechanism that only engages the shift
each N invokes of the function. So the larger the value, the longer
it will take to complete the fade.

Also if a new call is made to set a new value, that will merely
overwrite the new target colour and delay value, implying the new
setting will take effect even if the older coloiur had not yet
fully faded in.

Also added a new arg to force apmode by writing a marker to config
and rebooting which then forces ap mode. Handy when you are too
lazy to go over and power-cycle a device

Also major code cleanup in places and remodelled the in-memory
arrays now into indented structs to make it easier to manage