rework with support for manual switch detection where we buy ourselves an hour of bypass from managed timing.
So we can turn on a light and avoid having it turned
off automatically for an hour. Similarly turning off
devices will have the same effect.

It's still a work in progress as are the terms being
used to describe the switch terminologies.

I have work to do to integrate motion detection
via SR501 and will tweak as required when making those
changes