Removed WiFI.disconnect() call used right before the mode is set to STA
somehow this has an inherent issue with some devices
ending up in a limbo state unable to connect. Whatever
was the original reason for putting that call there, it
was a bum steer.

Also removed the STA reconnect code after N checks while
down. We now only apply a reboot after clocking 1800
checks for wifi up while in a known down state. That's
1 hour tolerance of outage before we reboot