Fixed WiFI recovery issue after outage of router. This was solved by getting the wifi up check function to restart STA mode after each 30 attempts.. roughly every 60 seconds. In that restart or STA mode, we added a call to WiFi.mode(WIFI_OFF) which will seemingly work around the glitch that prevents the existing STA mode from reconnecting.
The JSON status now has a "compile_date" field that gets
a hackjob of a Software Version from an in-memory string
whch was initialised from __DATE__ and __LINE__ C preprocessor
variables.. told you it was hacky. It will help me with
knowing how up to date a device is versus another. Will
replace this with a better solution eventually.

Added a function to restore the WiFI LED to the correct
state after we've completed the initial WiFI conect or
a later re-connect sequence. It sets the LED off initially
and then searches for a switch which is assigned that WiFI LED.
If found, the LED state is matched to the switch state.
Ensures we're restoring the LED to it's original state
after we used it for status reporting.