Added first stab at /json branch for webserver
The /json hook on the webserver will now
return a JSON document showing all discovered
device data. This is essentially a top-level JSON
document with a "devices" attribute that is a list
of each of the JSON status we have cached from the
probes performed on each device.

In addition to this, the URL can take args
for device, zone & control and be used to
turn on/off switches as required.

For example:
http://ip:8080/json?device=ESP-009E9496-Office&zone=Office&control=KRK+Monitors&state=1

will set the state of the 'KRK Monitors' to 1 for the Office Zone

At the moment, you need to also specify the device name. I'm tempted to
simplify this to Zone and Control only in time.

I will also think and rework the JSON response into something
more simplified, possibly showing only zones, devices and sensors.

The full status may be reserved for a special arg etc.

This next step is to be able to use the web server as a central
command point for other integrations with IFTTT, Google Home
and all that stuff