Added support for JSON config file
Config is now loaded from file ~/.jbhasd_web_server

On startup, the script tries to load this file
and parse contents as JSON into a config dict.
If this fails, it initialises a default config
dict and writes that to the config file.
So on first run, it will create the file with
defaults.

Then where we previously had various global
variables being accessed, we now do direct
dictionary lookups from the config dict.