Reworked zone and device consoles refresh/action behaviour
Changed the javascript driven click actions to now call
an action URL to perform the desired on/off action direct
to a /api?xxx=yy URL. The same click code also then calls the
root /zone or /device URL to get an instant update on data
for it to live update the screen.

This separates API calls to webserver from calls to refresh
and ensuring that all API activity to the server is truly
separate from GUI-based presentation.