Significant rework of the design to cater for the new model.
The reset config function is now
dedicated to the sole purpose of defining
default setings and comitting these to EEPROM

Load config will default to initialising the internal
device config struct and related lists as well as parsing
from the JSON it retrieved from EEPROM.

If the load stage encounters and error with data, it
invokes a reset_config which will apply sane defaults and
commit to disk.

This new flow is better all round for a more
progressive model.

LED work still not complete.

Of note here is the new provisioned field. This set to 0
when a reset_config is called. When someone provisions the
config via JSON, this field is set to 1. So the device
will get online after it's initial AP Mode config where it
has wifi and zone set. Thereafter, it will activate on the LAN
with telnet logging, MDNS etc. But discovery can leverage this
field as being 0 to trigger the auto-provisioning of an
assigned configuration profile. More to come on this
as it evolves