Several cleanups
Cleaned up all global defs to place one per line

Modified Config load to operate in its own thread
and do a reload check every 5 seconds based on timestamp
of config file. If the load fails (json parse), then it
continues with the existing config. This allows a better
run model with on-the-fly edits to config

Changed time format in timers to be HH:MM and got code to
conveniently remove the ':' when processing. This keeps
the handy numeric aspect to times while letting us use
a valid format in config items