Added support for Addressable RGB strips.
The model is similar to RGB support in that a program
string is configured against the given control and used to
drive a pattern sequence.

This is somewhat harder than the RGB strip as the
variables are more complex. So the initial feature
set is a sequence of colours that can be displayed and
progressed through the strip in formward or backward
direction and some control how the strip is cleared
and how the drawing progresses. More thought will be
needed here if something more sophisticated is required