Cleanup of all int 1,2,4 octet types from usigned char/short/int into uintX_t variations purely for readibility and to update to more modern norms. As many fields were reduced to single octet sizes as possible.
Changed LOW/HIGH handling of LED and relay GPIO
to react to XXX_on_high and XXX_on_high fields
that dictate the ON state to be HIGH. These fields
are paired with the WiFI LED pin and switch LED and
relay pins so that in all cases we are capable of
configuring the ON state to be LOW or HIGH as required.
I left toggle button support for the boot pin and
switch manual to be LOW for now. Can be augmented the
same way in the future if required.

Changed the AP mode settings screen to only allow it
refresh the WiFI scan data evry 30 seconds and once on
startup. That prevents it stalling when refreshed.

Added a Reset Config option to the AP screen to force
a reset of config. This used to be done via an empty
SSID field which is no longer possible given the SSID
is now a combo box.