Updated loop_task_check_wifi_up() internals to handle newer 10-second call interval (versus older period of 2 seconds).
The max checks before rboot is not 360.. giving it an hour
grace before reboot and the wifi restart count is 6 giving it a
1 minute before wifi restart.

Honestly I'd love to refactor this into something that tracks
timestamps (millis()) and works with elapsed intervals to determine
the outage time. Then I could call it more/less frequently abd still
get the same outcome. Maybe sometime :)