Added first stab at motion control for switches
we can now define a pin used to detect motion
and use it to activate a switch.

The config sets "sw_motion_pin" in the switch
definition to the desired pin and also includes
a value for "sw_motion_interval" (seconds) to define
the interval to apply when motion is detected

The reported switch context in the JSON status
will be "motion".

the WEB device API can dynamically set this motion
interval to 0 (0ff) or some value in seconds as a
means of activating motion control as it will need to
be timed and disabled for example during daytime