Slight rework on how motion detection intervals are managed.
When motion is now detected, we will always update the last activity
msec timestamp even if the switch is already on.

We will only turn the switch on if its in the off state
when motion is detected and related to that, we will only
push an update notification to the server if we actually
turned on the switch.

So this will result in the PIR pin causing multiple motion events
as it retains a HIGH status on the pin. The effect of this
is that we have to observe an idle period of the motion interval
seconds before we will turn off the switch. So if it gets triggered
during the interval, the on state is carried over until we
measure no more motion for the defined motion interval.

The result is that motion turns on the switch and keeps it on
as long as motion is still being detected. In the earlier release
it came on when motion was detected and then turned off exactly
motion_interval seconds later and would come straight on again if
motion was still in play. That just made for an undesirable
experience.