Rework of timer for Motion Support
Added sunset keyword for times

Reworked parsing of sunset and sunrise into common function

Enhanced check_switch function to now check for motion_interval
and return it along with the existing desired_state. The intent
was to enhance the forced switch on/off call to include the
motion_interval.

Timers now have a motion_on, motion_off time and a motion_interval
value allowing the user to define a motion time period and
timeout interval when it triggers