Rework of AP Mode Wifi Scan into a manual scan be means of a form button. The scan is also performed initially on the first run. Provides a more responsive experience when trying to select an SSID that did not appear in the most recent scan
In wifi_init(), added a first run static to limit it's calls to a once
off only. Nothing were re-entrant unsafe but it makes for cleaner
logging.

Removed WiFi.persistent(false) as it was not doing what I had
asumed it did.

Also removed WiFi.mode(WIFI_OFF) and added in call to WiFi.disconnect()
in the STA mode start function. Helping to remove resets occurring when
handling an wifi outage