Added support for idle status detection based around calls to the /status API function.
The principle is to track the timestamp of the most recent call to
/status and then detect apparent idle scenarios when this timestamp
is not being updated.

Two idle periods are defined in the top-level JSON
config:

* "idle_period_wifi"
Defines the interval in seconds since the last /status call that
triggers a restart of WiFi

* "idle_period_reboot"
Defines the interval in seconds since the last /data call that
triggers a reboot of the device

You can set either or both of these periods depending on needs. Each is
disabled if omitted from JSON config or set to 0.

The reason for this feature was based on observing devices in my home
mesh network.

The observation was that the monitoring web server starts timing out /status
calls to some devices and eventually scrubs the devices from it's lists.
It's not the same devices all the time but can sometimes occur more
frequently wity one specific device versus another.

After being scrubbed from the lists, the devices then fail to be rediscovered
either indefinitely or for long periods. As a result, the home automation
and network-based control is lost.

The odd thing is that the devices continue to work fine responding to
manual button presses. They do not show any status LED flashing to
suggest they think they have lost WiFi. Yet when you power-cycle the
nearest WiFi router, the devices immediately react by entering the WiFi
reconnect state and start to flash the status LED. So it seems that while
they are not accessible or discoverable on the LAN, they are consciously
aware that they are connected to WiFi.

I suspect the underlying issue is a bug with the WiFi routers losing LAN
access to the devices. Maybe it is specific to the 8266 code or
hardware.

But in any event, it posed a challenge to create a work-around and I felt
the best was was to build a health check around the monitoring of calls
to /status as that is the exact thing the webserver uses to know the
given device is still alive and kicking.

If the device fails to see this /status function being called, it will
react by restarting WiFi every "idle_period_wifi" seconds or reboot after
the larger "idle_period_reboot" seconds.

Logical values for these would be to set "idle_period_wifi" to 300 (5 minutes)
and "idle_period_reboot" to 7200 (2 hours)