modified MDNS setup to skip any checks on success and also call a MDNS.close() before it does the begin as it makes it correctly restart the service.
This is ongoing investigation into why MDNS seems to be malfunctioning
in some form. It seems to stop working after a while for some devices
some of the time. The idle status timer which does a restart of WiFi and
related serves seems to rectify the issue. Still needs more work to get
a better solution.