Discovery/purge reworks
Changed discovery to store current time in status ts dict as each device is
discovered. This gives us a grace period of the standard purge time for
each device from discovery to the first valid store of a status value.

That change then deprecated the need to detect devices we had never
corectly probed.

Also removed the change detection based on diffing the controls data as
this is not invalid given time deltas are being reported and they will
always be changing