fix to force a switch off when the motion interval is set to 0 while the switch is in a motion context.
Fixes an edge case where the switch enters a motion scenario right
before an automated event sets the motion_interval to 0.

The loop_task_check_switches() function designed to check for manual
presses and clear down of motion scenarios will not clear down that
existing motion scenario because of the interval being 0.