Added RGB device init interval to status response.
Disabled WiFi persistence (writing/reading previous settings from flash on start). Turned
out to be a red herring change as it related to a slight dimming on H801
devices and was assumed to relate to flash writes WRT WiFi module
activities. In any event, it is no harm as the app uses its own config
to manage WiFi credentials.