rework of event time into optional times array to allow the same event be used for a list of times
Also modified the event interval to 60 seconds grace ensuring that we
have per-minute granularity on events.

Do need to revisit this so that we can program the most recent event on
any device after it is discovered (or rediscovered). This will help
retain some level of continuity but needs a rethink again on manual
over-rides and boot up periods on devices that would want to retain a
certain fix on their current states/programs