Added a sleep() method to TaskMan class to allow it determine a sleep period in msecs based on the next tak to fire and then calls delay() for that number of msecs. It protects itself from sleeping over 1000 msecs. This call to TaskMan:sleep() would be placed in the main loop() function after the nudge() call.
This dramatically reduces the agression of the nudge() function in
finding tasks to execute on each call of loop(). Observations already
show that some devices heat less as a result (observed on a Sonoff S26R2 esp8255)

For tasks scheduled at 1 msec intervals, such as those for RGB and aRGB
transitions, the sleep skips the delay and returns. This then negates
the efficiency gained. More work needed here to try and dynamically
detect RGB/aRGB program transitions being active to then control the use
of the sleep bypass.